Refactor the Payments API
