+++ mollie/api/resources/__init__.py
@@ -1,7 +1,9 @@
+from .payments import Payments
 from .refunds import OrderRefunds, PaymentRefunds, Refunds, SettlementRefunds
 
 __all__ = [
     OrderRefunds,
+    Payments,
     PaymentRefunds,
     Refunds,
     SettlementRefunds,
