+++ tests/test_api_client.py
@@ -144,7 +144,6 @@ def test_client_invalid_update_data(client):
     "endpoint, errorstr",
     [
         ("customers", "Invalid customer ID: 'invalid'. A customer ID should start with 'cst_'."),
-        ("payments", "Invalid payment ID: 'invalid'. A payment ID should start with 'tr_'."),
         ("orders", "Invalid order ID: 'invalid'. An order ID should start with 'ord_'."),
         ("captures", "Invalid capture ID: 'invalid'. A capture ID should start with 'cpt_'."),
         ("invoices", "Invalid invoice ID: 'invalid'. An invoice ID should start with 'inv_'."),
@@ -166,7 +165,6 @@ def test_client_get_invalid_id(client, endpoint, errorstr):
     "endpoint, errorstr",
     [
         ("customer_mandates", "Invalid mandate ID: 'invalid'. A mandate ID should start with 'mdt_'."),
-        ("customer_payments", "Invalid payment ID: 'invalid'. A payment ID should start with 'tr_'."),
         ("customer_subscriptions", "Invalid subscription ID: 'invalid'. A subscription ID should start with 'sub_'."),
     ],
 )
